responsible-ai-web-app-main/lyrics_gen_api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "responsible_lyrics_gen"
version = "1.0.0"
authors = [
    { name = "Winand Metz", email = "winand.metz@hva.nl" },
    { name = "Caleigh Esterhuyse", email = "caleigh.esterhuyse@hva.nl" }
]
description = "A package which holds tools usefull for generative AI text generation. In this case dutch \"smartlap lyrics\""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lingua-language-detector>=1.3.2",
    "wheel>=0.40.0",
    "responsible_generative_ai>=1.0.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Source" = "https://gitlab.fdmci.hva.nl/metzwm/responsible-ai-gpt2-proto"
"Bug Tracker" = "https://gitlab.fdmci.hva.nl/metzwm/responsible-ai-gpt2-proto/-/issues"

[tool.hatch.build]
ignore-vcs = true
